+++ src/xyz/oihana/schema/constants/traits/AppointmentsTrait.php
@@ -3,6 +3,7 @@
 namespace xyz\oihana\schema\constants\traits;
 
 use xyz\oihana\schema\constants\traits\appointments\AppointmentTrait;
+use xyz\oihana\schema\constants\traits\appointments\FollowUpTrait;
 use xyz\oihana\schema\constants\traits\appointments\VisitReportTrait;
 
 /**
@@ -15,5 +16,6 @@ use xyz\oihana\schema\constants\traits\appointments\VisitReportTrait;
 trait AppointmentsTrait
 {
     use AppointmentTrait ,
+        FollowUpTrait    ,
         VisitReportTrait ;
 }