+++ src/xyz/oihana/schema/constants/traits/appointments/FollowUpTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\appointments;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\appointments\FollowUp} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\appointments
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.5.0
+ */
+trait FollowUpTrait
+{
+    const string FOLLOW_UP      = 'followUp'     ;
+    const string FOLLOW_UP_TYPE = 'followUpType' ;
+}